+++ src/index.ts
@@ -1,0 +1,18 @@
+// Component exports
+export { default as Button } from './Button'
+export type { ButtonProps } from './Button'
+
+// TODO: Add additional component exports here as they are converted to TypeScript
+// Example:
+// export { default as Col } from './Col'
+// export type { ColProps } from './Col'
+
+// For backwards compatibility
+import Button from './Button'
+
+// TODO: Add additional components here
+const Harmonium = {
+  Button,
+}
+
+export default Harmonium